feature #32 added continue
