+++ blackjack/part_class/game_result.rb
