+++ main.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative "./blackjack/blackjack.rb"
+
+Blackjack.new
